Working kmer tokenizer
This is a bit more involved - it uses word piece because BPE just goes for the smallest token without a merges file.

We require pre-tokenization (which is actually kmerisation) and a set of tail tokens down to 1-mers

This is also now returning a non-fast tokenizer. Not that I think it matters, but it is different and I should standardise the others to this as well